CLI: Use stdout and stderr to print Messages with xmlschema-validate command (#393)
* feat(cli-validate): Use stdout and stderr instead of print statement

This change helps users to distinguish between error messages and
others, e.g. in CI pipelines.

Signed-off-by: Sven Sager <s.sager@kunbus.com>

* feat(cli-validate): Use argument -v to print errors on stderr

The argument -v (verbosity) is not yet used in the validate function.
The specification of -v is now used in xmlschema-validate to obtain a
detailed error report on the stderr. This helps users to find the error
in their XML files at the mentioned place.

Signed-off-by: Sven Sager <s.sager@kunbus.com>

---------

Signed-off-by: Sven Sager <s.sager@kunbus.com>